Добавил Entity таблиц Location, Category и Event
