--- Homework/task5_6/src/main/java/org/yakdanol/task5_6/model/Category.java
@@ -1,16 +1,0 @@
-package org.yakdanol.task5_6.model;
-
-import lombok.*;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Data
-public class Category {
-    private Long id;
-
-    @NonNull
-    private String slug;
-
-    @NonNull
-    private String name;
-}
