+++ Homework/task5_6/src/main/java/org/yakdanol/task5_6/model/entity/Category.java
@@ -1,0 +1,24 @@
+package org.yakdanol.task5_6.model.entity;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Table(name = "categories")
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class Category {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false)
+    private String slug;
+
+    @Column(nullable = false)
+    private String name;
+}
